refactor: add index.d.ts and update declare module
