+++ packages/axios/index.d.ts
@@ -1,9 +1,9 @@
-import { I18nOptions } from './dist/';
+import { AxiosRequestConfig } from 'axios';
 
 export * from './dist/index';
 
 declare module '@midwayjs/core' {
   interface MidwayConfig {
-    i18n?: PowerPartial<I18nOptions>;
+    axios?: AxiosRequestConfig;
   }
 }
