+++ packages/axios/src/interface.ts
@@ -47,10 +47,3 @@ export interface AxiosHttpService {
     config?: AxiosRequestConfig
   ): Promise<R>;
 }
-
-declare module '@midwayjs/core/dist/interface' {
-  interface MidwayConfig {
-    axios?: AxiosRequestConfig;
-  }
-}
-
