--- packages/cache/.gitignore
@@ -1,15 +1,0 @@
-logs/
-npm-debug.log
-yarn-error.log
-node_modules/
-package-lock.json
-yarn.lock
-coverage/
-dist/
-.idea/
-run/
-.DS_Store
-*.sw*
-*.un~
-.tsbuildinfo
-.tsbuildinfo.*
