+++ packages/cache/index.d.ts
@@ -2,7 +2,7 @@ import { CacheOptions, StoreConfig } from 'cache-manager';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     cache?: StoreConfig & CacheOptions;
   }
