+++ packages/cos/index.d.ts
@@ -2,7 +2,7 @@ import * as COS from 'cos-nodejs-sdk-v5';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     cos?: ServiceFactoryConfigOption<COS.COSOptions>;
   }
