+++ packages/cross-domain/index.d.ts
@@ -1,7 +1,7 @@
 import { CORSOptions, JSONPOptions } from './dist/index';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     cors?: Partial<CORSOptions>;
     jsonp?: Partial<JSONPOptions>;
