+++ packages/express-session/index.d.ts
@@ -1,7 +1,7 @@
 import { SessionOptions } from 'express-session';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     session?: Partial<
       SessionOptions & {
