+++ packages/grpc/index.d.ts
@@ -1,0 +1,10 @@
+import { DefaultConfig, IMidwayGRPFrameworkOptions } from './dist';
+
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    grpcServer?: IMidwayGRPFrameworkOptions;
+    grpc?: PowerPartial<DefaultConfig>;
+  }
+}
