+++ packages/grpc/package.json
@@ -3,7 +3,7 @@
   "version": "3.1.6",
   "description": "Midway Framework for gRPC",
   "main": "dist/index",
-  "typings": "dist/index.d.ts",
+  "typings": "index.d.ts",
   "scripts": {
     "build": "tsc",
     "test": "node --require=ts-node/register ../../node_modules/.bin/jest --runInBand",
@@ -15,11 +15,12 @@
     "IoC",
     "web",
     "scene",
-    "express"
+    "grpc"
   ],
   "files": [
     "dist/**/*.js",
-    "dist/**/*.d.ts"
+    "dist/**/*.d.ts",
+    "index.d.ts"
   ],
   "license": "MIT",
   "devDependencies": {
