+++ packages/http-proxy/index.d.ts
@@ -1,7 +1,7 @@
 import { HttpProxyConfig } from './dist/index';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     httpProxy?: Partial<HttpProxyConfig> | Partial<HttpProxyConfig>[];
   }
