+++ packages/info/index.d.ts
@@ -1,7 +1,7 @@
 import { InfoConfigOptions } from './dist/index';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     info?: Partial<InfoConfigOptions>;
   }
