+++ packages/jwt/index.d.ts
@@ -2,7 +2,7 @@ import { SignOptions } from 'jsonwebtoken';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     jwt?: SignOptions & {
       secret?: string;
