+++ packages/mongoose/index.d.ts
@@ -1,0 +1,11 @@
+import { ConnectionOptions } from './dist';
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    mongoose?: ServiceFactoryConfigOption<{
+      uri: string;
+      options: ConnectionOptions;
+    }>;
+  }
+}
