+++ packages/mongoose/package.json
@@ -3,7 +3,7 @@
   "version": "3.1.6",
   "description": "Midway Component for mongoose",
   "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
+  "typings": "index.d.ts",
   "scripts": {
     "build": "tsc",
     "test": "node --require=ts-node/register ../../node_modules/.bin/jest --runInBand",
@@ -14,7 +14,8 @@
   "author": "czy88840616@gmail.com",
   "files": [
     "dist/**/*.js",
-    "dist/**/*.d.ts"
+    "dist/**/*.d.ts",
+    "index.d.ts"
   ],
   "engines": {
     "node": ">=12"
