+++ packages/orm/index.d.ts
@@ -2,7 +2,7 @@ import { ConnectionOptions } from './dist/index';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     orm?:
       | Partial<ConnectionOptions>
