+++ packages/oss/index.d.ts
@@ -2,7 +2,7 @@ import * as OSS from 'ali-oss';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     oss?: ServiceFactoryConfigOption<OSS.STSOptions | OSS.Options>;
   }
