+++ packages/otel/index.d.ts
@@ -1,6 +1,6 @@
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     otel?: Record<string, any>;
   }
