+++ packages/passport/index.d.ts
@@ -2,7 +2,7 @@ import * as passport from 'passport';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     passport?: passport.AuthenticateOptions;
   }
