+++ packages/rabbitmq/index.d.ts
@@ -1,0 +1,8 @@
+import { IMidwayRabbitMQConfigurationOptions } from './dist';
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    rabbitmq?: PowerPartial<IMidwayRabbitMQConfigurationOptions>;
+  }
+}
