+++ packages/rabbitmq/package.json
@@ -3,7 +3,7 @@
   "version": "3.1.6",
   "description": "Midway Framework for rabbitmq",
   "main": "dist/index",
-  "typings": "dist/index.d.ts",
+  "typings": "index.d.ts",
   "scripts": {
     "build": "tsc",
     "test": "node --require=ts-node/register ../../node_modules/.bin/jest --runInBand",
@@ -19,7 +19,8 @@
   ],
   "files": [
     "dist/**/*.js",
-    "dist/**/*.d.ts"
+    "dist/**/*.d.ts",
+    "index.d.ts"
   ],
   "license": "MIT",
   "devDependencies": {
