+++ packages/sequelize/index.d.ts
@@ -2,7 +2,7 @@ import { SequelizeOptions } from 'sequelize-typescript';
 
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     sequelize?: {
       options?: SequelizeOptions;
