+++ packages/socketio/index.d.ts
@@ -1,0 +1,8 @@
+import { IMidwaySocketIOConfigurationOptions } from './dist';
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    socketIO?: IMidwaySocketIOConfigurationOptions;
+  }
+}
