+++ packages/static-file/index.d.ts
@@ -1,7 +1,7 @@
 import { StaticFileOptions } from './dist/index';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     staticFile?: Partial<StaticFileOptions>;
   }
