+++ packages/tablestore/index.d.ts
@@ -1,0 +1,8 @@
+import { TableStoreConfig } from './dist';
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    tableStore?: ServiceFactoryConfigOption<TableStoreConfig>;
+  }
+}
