+++ packages/tablestore/src/interface.ts
@@ -840,9 +840,3 @@ export interface TableStoreConfig {
    */
   computeChecksums?: boolean;
 }
-
-declare module '@midwayjs/core/dist/interface' {
-  interface MidwayConfig {
-    tableStore?: ServiceFactoryConfigOption<TableStoreConfig>;
-  }
-}
