+++ packages/upload/index.d.ts
@@ -1,7 +1,7 @@
 import { UploadFileInfo, UploadOptions } from './dist/index';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     upload?: Partial<UploadOptions>;
   }
