+++ packages/validate/index.d.ts
@@ -1,7 +1,7 @@
 export * from './dist/index';
 import * as Joi from 'joi';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     validate?: {
       validationOptions?: Joi.ValidationOptions;
