+++ packages/view-nunjucks/index.d.ts
@@ -1,7 +1,7 @@
 import _default from './dist/config/config.default';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   // eslint-disable-next-line
   interface MidwayConfig extends Partial<typeof _default> {}
 }
