+++ packages/view/index.d.ts
@@ -1,7 +1,7 @@
 import { IViewEngine } from './dist';
 export * from './dist/index';
 
-declare module '@midwayjs/core/dist/interface' {
+declare module '@midwayjs/core' {
   interface MidwayConfig {
     view?: {
       /**
