+++ packages/web/package.json
@@ -3,7 +3,7 @@
   "version": "3.1.6",
   "description": "Midway Web Framework for Egg.js",
   "main": "dist/index",
-  "typings": "dist/index.d.ts",
+  "typings": "index.d.ts",
   "scripts": {
     "build": "tsc",
     "test": "node --require=ts-node/register ../../node_modules/.bin/jest --runInBand",
@@ -23,7 +23,8 @@
     "agent.js",
     "app.js",
     "dist/**/*.js",
-    "dist/**/*.d.ts"
+    "dist/**/*.d.ts",
+    "index.d.ts"
   ],
   "license": "MIT",
   "devDependencies": {
