+++ packages/web/src/interface.ts
@@ -1,10 +1,8 @@
-import { Context as EggContext, Application as EggApplication, EggAppConfig } from 'egg';
+import { Context as EggContext, Application as EggApplication } from 'egg';
 import {
   IMidwayContainer,
   IMidwayContext,
-  Context as IMidwayBaseContext,
   IMidwayApplication,
-  IMidwayBaseApplication,
   IConfigurationOptions,
   NextFunction as BaseNextFunction,
 } from '@midwayjs/core';
@@ -17,24 +15,6 @@ export interface IMidwayWebBaseApplication {
   getCoreLogger(): ILogger;
   generateMiddleware?(middlewareId: any): Promise<Middleware<DefaultState, EggContext>>;
   createLogger(name: string, options: LoggerOptions): ILogger;
-}
-
-declare module 'egg' {
-  interface EggAppInfo {
-    appDir: string;
-  }
-
-  // 这里再次覆盖和 egg 不同的定义，不然 egg 插件里可能会报错
-  interface Application extends IMidwayBaseApplication<Context>, IMidwayWebBaseApplication {
-    createAnonymousContext(...args: any[]): EggContext;
-    getCoreLogger(): EggLogger & ILogger;
-    getLogger(name?: string): EggLogger & ILogger;
-    createLogger(name: string, options: LoggerOptions): EggLogger & ILogger;
-  }
-
-  interface Context <ResponseBodyT = any> extends IMidwayBaseContext {
-    getLogger(name?: string): ILogger;
-  }
 }
 
 export type IMidwayWebApplication = IMidwayApplication<Context, EggApplication & IMidwayWebBaseApplication>;
@@ -98,10 +78,4 @@ export type MidwayWebMiddleware = Middleware<DefaultState, Context>;
  */
 export interface IWebMiddleware {
   resolve(): MidwayWebMiddleware;
-}
-
-declare module '@midwayjs/core/dist/interface' {
-  interface MidwayConfig extends PowerPartial<EggAppConfig> {
-    egg?: IMidwayWebConfigurationOptions;
-  }
 }
