+++ packages/ws/index.d.ts
@@ -1,0 +1,9 @@
+import { IMidwayWSConfigurationOptions } from './dist';
+
+export * from './dist/index';
+
+declare module '@midwayjs/core' {
+  interface MidwayConfig {
+    webSocket?: IMidwayWSConfigurationOptions;
+  }
+}
