+++ packages/ws/src/interface.ts
@@ -20,9 +20,3 @@ export type IMidwayWSContext = IMidwayContext<WebSocket & {
 export type Application = IMidwayWSApplication;
 export type NextFunction = BaseNextFunction;
 export interface Context extends IMidwayWSContext {}
-
-declare module '@midwayjs/core/dist/interface' {
-  interface MidwayConfig {
-    webSocket?: IMidwayWSConfigurationOptions;
-  }
-}
